Rename consts module to constants
