+++ tests/unit/test_flask_app.py
@@ -7,7 +7,7 @@ import json
 import pytest
 
 from charm_state import CharmState
-from consts import FLASK_ENV_CONFIG_PREFIX
+from constants import FLASK_ENV_CONFIG_PREFIX
 from flask_app import FlaskApp
